% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{CalculateGDS}
\alias{CalculateGDS}
\title{Calculate the Global Discriminant Score (GDS)}
\usage{
CalculateGDS(clusters.DA = NULL)
}
\arguments{
\item{clusters.DA}{Data frame with the output from the DiscriminantAnalysis() function.}
}
\value{
Data frame with the GDS for each clustering resolution (k).
}
\description{
This function calculates the average maximum discriminancy (for all classes) at each clustering resolution (k).
}
