% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{DiscriminantAnalysis.plus.seurat}
\alias{DiscriminantAnalysis.plus.seurat}
\title{Discriminant Analysis Plus Seurat}
\usage{
DiscriminantAnalysis.plus.seurat(
  seurat.object = NULL,
  assay = NULL,
  slot = "data",
  annotation = NULL,
  features = NULL,
  BackgroundClasses = NULL,
  QueryClasses = NULL,
  reduction.name = NULL,
  pairwise = NULL
)
}
\arguments{
\item{seurat.object}{Seurat object}

\item{assay}{Name of the assay in \code{seurat.object} to pull the slot data from. If NULL, it takes the name of the active assay in \code{seurat.object}, default = NULL.}

\item{slot}{Slot to pull data from, default: 'data'.}

\item{annotation}{Character vector with the class of the samples.}

\item{features}{Character vector with the names of the features to be included in the discriminant analysis. If NULL, all the features in data rows are included. Default: NULL}

\item{BackgroundClasses}{Classes to compare with the query class. If NULL, all the classes in annotation except the query class will be included as background. Default: NULL}

\item{QueryClasses}{Character vector with the classes to compare with respect to the background. If NULL, all the classes in annotation will be analyzed (separately) with respect to the background. Default: NULL}

\item{reduction.name}{Name to label the analysis. Default: NULL}

\item{pairwise}{Logical. If TRUE, a pairwise comparison between classes is performed. If NULL, each query class is compared with respect to the background classes. Default: NULL}
}
\value{
The Seurat object including a list with complete results from the \code{ropls::opls()}.
}
\description{
This function computes the discriminant analysis (OPLS-DA from the \code{ropls} library), comparing every class in a categorical variable against the background (i.e., the other classes). It provides more information than DiscriminantAnalysis(), as it includes the complete output from the \code{ropls::opls()} function.
}
\seealso{
\code{\link[Seurat]{reexports}}
}
