% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-ModulonR_functions_wo_comments.R
\name{DiscriminantAnalysis_Plus_Seurat}
\alias{DiscriminantAnalysis_Plus_Seurat}
\title{FUNCTION_TITLE}
\usage{
DiscriminantAnalysis_Plus_Seurat(
  seurat.object = NULL,
  assay = NULL,
  slot = "data",
  annotation = NULL,
  features = NULL,
  BackgroundClasses = NULL,
  QueryClasses = NULL,
  reduction.name = NULL,
  pairwise = NULL
)
}
\arguments{
\item{seurat.object}{PARAM_DESCRIPTION, Default: NULL}

\item{assay}{PARAM_DESCRIPTION, Default: NULL}

\item{slot}{PARAM_DESCRIPTION, Default: 'data'}

\item{annotation}{PARAM_DESCRIPTION, Default: NULL}

\item{features}{PARAM_DESCRIPTION, Default: NULL}

\item{BackgroundClasses}{PARAM_DESCRIPTION, Default: NULL}

\item{QueryClasses}{PARAM_DESCRIPTION, Default: NULL}

\item{reduction.name}{PARAM_DESCRIPTION, Default: NULL}

\item{pairwise}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Seurat]{reexports}}
}
