% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{FindMarkersOPLS.ext.wo.seurat}
\alias{FindMarkersOPLS.ext.wo.seurat}
\title{FUNCTION_TITLE}
\usage{
FindMarkersOPLS.ext.wo.seurat(
  data,
  annotation,
  ident.1,
  ident.2 = NULL,
  features = NULL,
  ...
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{annotation}{PARAM_DESCRIPTION}

\item{ident.1}{PARAM_DESCRIPTION}

\item{ident.2}{PARAM_DESCRIPTION, Default: NULL}

\item{features}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ropls]{opls}}
}
