% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{FindMarkersOPLS}
\alias{FindMarkersOPLS}
\title{FUNCTION_TITLE}
\usage{
FindMarkersOPLS(
  mat,
  annotation,
  ident.1,
  ident.2 = NULL,
  scale_weights = TRUE,
  scale_vipVn = TRUE,
  ...
)
}
\arguments{
\item{mat}{PARAM_DESCRIPTION}

\item{annotation}{PARAM_DESCRIPTION}

\item{ident.1}{PARAM_DESCRIPTION}

\item{ident.2}{PARAM_DESCRIPTION, Default: NULL}

\item{scale_weights}{PARAM_DESCRIPTION, Default: TRUE}

\item{scale_vipVn}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ropls]{opls}}
}
