% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{GetHierarchicalClusters}
\alias{GetHierarchicalClusters}
\title{Get Hierarchical Clusters}
\usage{
GetHierarchicalClusters(
  mat,
  distance.method = DISTANCE_METHODS,
  clustering.method = CLUSTERING_METHODS,
  scale = TRUE,
  ...
)
}
\arguments{
\item{mat}{The input matrix for which the features in the rows will be clustered.}

\item{distance.method}{The distance method to use in \code{stats::dist()}.}

\item{clustering.method}{The hierarchical clustering method to use in \code{stats::hclust()}. Default is "complete" (see ?stats::hclust for more information).}

\item{scale}{Logical argument indicating whether the rows of the input matrix should be scaled (mean-centered and divided by the standard deviation). Default is TRUE.}

\item{...}{Additional parameters for \code{stats::hclust}.}
}
\value{
An object of class hclust (see ?stats::hclust).
}
\description{
This function computes the hierarchical clustering tree of the features in the rows of a matrix
using the \code{stats::hclust()} method.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
  # Generate a random matrix with 100 rows and 5 columns
  set.seed(123)
  random_matrix <- matrix(abs(rnorm(25 * 5)), nrow = 25, ncol = 5)
  rownames(random_matrix) <- paste0("F", 1:25)
  plot(GetHierarchicalClusters(mat = random_matrix))
  }
}
}
