% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{ModulonIdent}
\alias{ModulonIdent}
\title{Modulon Identification}
\usage{
ModulonIdent(
  data,
  annotation,
  BackgroundClasses = NULL,
  QueryClasses = NULL,
  k.range = c(2, 30)
)
}
\arguments{
\item{data}{Matrix of regulon activity with features (regulons) in rows and samples (cells) in columns.}

\item{annotation}{Character vector with the class (phenotype/state) of the samples (cells).}

\item{BackgroundClasses}{Classes to compare with the query class. If NULL, all the classes in annotation except the query class will be included as background. Default: NULL}

\item{QueryClasses}{Character vector with the classes to compare with respect to the background. If NULL, all the classes in annotation will be analyzed (separately) with respect to the background. Default: NULL}

\item{k.range}{Clustering resolution (k) range to be explored. Default: c(2, 30)}
}
\value{
List with the identified modulons and the global discriminant score (GDS) at every k within the range defined by k.range.
}
\description{
Wrapper function to perform the first step of a modulon analysis: modulon identification.
}
