% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R,
%   R/oxy-ModulonR_functions_wo_comments.R
\name{ModulonPert}
\alias{ModulonPert}
\title{FUNCTION_TITLE}
\usage{
ModulonPert(
  Regulons,
  Modulons,
  ExpMat,
  annotation,
  BackgroundClasses = NULL,
  TargetState,
  TargetModulon,
  CombSize,
  Weights = NULL
)

ModulonPert(
  Regulons,
  Modulons,
  ExpMat,
  annotation,
  BackgroundClasses = NULL,
  TargetState,
  TargetModulon,
  CombSize,
  Weights = NULL
)
}
\arguments{
\item{Regulons}{PARAM_DESCRIPTION}

\item{Modulons}{PARAM_DESCRIPTION}

\item{ExpMat}{PARAM_DESCRIPTION}

\item{annotation}{PARAM_DESCRIPTION}

\item{BackgroundClasses}{PARAM_DESCRIPTION, Default: NULL}

\item{TargetState}{PARAM_DESCRIPTION}

\item{TargetModulon}{PARAM_DESCRIPTION}

\item{CombSize}{PARAM_DESCRIPTION}

\item{Weights}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
