% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR-package.R
\docType{package}
\name{ModulonR-package}
\alias{ModulonR}
\alias{ModulonR-package}
\title{ModulonR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
