% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R,
%   R/oxy-ModulonR_functions_wo_comments.R
\name{range01}
\alias{range01}
\title{FUNCTION_TITLE}
\usage{
range01(x)

range01(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
