% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{RunOPLS.plus}
\alias{RunOPLS.plus}
\title{Run OPLS Plus}
\usage{
RunOPLS.plus(data, annotation, ident.1, ident.2 = NULL, features = NULL, ...)
}
\arguments{
\item{data}{Matrix with features and samples in rows and columns, respectively.}

\item{annotation}{Character vector with the class of the samples.}

\item{ident.1}{Character, query class.}

\item{ident.2}{Character vector with background classes. If NULL, all the classes in annotation but the query class will be included, Default: NULL.}

\item{features}{Features to be included in the analysis. If NULL, all the features in in the assay and slot are included , Default: NULL.}
}
\value{
List with complete results from the \code{ropls::opls()} function for all comparisons.
}
\description{
Wrapper function to run the ropls::opls().
}
\seealso{
\code{\link[ropls]{opls}}
}
