% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{RunOPLS.plus.seurat}
\alias{RunOPLS.plus.seurat}
\title{FUNCTION_TITLE}
\usage{
RunOPLS.plus.seurat(
  seurat.object,
  ident.1,
  ident.2 = NULL,
  features = NULL,
  assay = NULL,
  slot = "data",
  ...
)
}
\arguments{
\item{seurat.object}{PARAM_DESCRIPTION}

\item{ident.1}{PARAM_DESCRIPTION}

\item{ident.2}{PARAM_DESCRIPTION, Default: NULL}

\item{features}{PARAM_DESCRIPTION, Default: NULL}

\item{assay}{PARAM_DESCRIPTION, Default: NULL}

\item{slot}{PARAM_DESCRIPTION, Default: 'data'}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Seurat]{reexports}}
\code{\link[ropls]{opls}}
}
