% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{RunOPLSDA}
\alias{RunOPLSDA}
\title{FUNCTION_TITLE}
\usage{
RunOPLSDA(
  data = NULL,
  annotation = NULL,
  classes = NULL,
  features = NULL,
  reduction.name = NULL,
  pairwise = NULL
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION, Default: NULL}

\item{annotation}{PARAM_DESCRIPTION, Default: NULL}

\item{classes}{PARAM_DESCRIPTION, Default: NULL}

\item{features}{PARAM_DESCRIPTION, Default: NULL}

\item{reduction.name}{PARAM_DESCRIPTION, Default: NULL}

\item{pairwise}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
