% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{scale_range}
\alias{scale_range}
\title{scale_range}
\usage{
scale_range(x)
}
\arguments{
\item{x}{Numerical vector.}
}
\value{
Numerical vector. The input numerical vector scaled to range from 0 to 1.
}
\description{
This function scales a numerical vector from 0 to 1.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 scale_range(x = c(0.1, 0.2, 0.1, 0.3))
 }
}
}
