% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R,
%   R/oxy-ModulonR_functions_wo_comments.R
\name{CalculateGDS}
\alias{CalculateGDS}
\title{FUNCTION_TITLE}
\usage{
CalculateGDS(clusters.DA = NULL)

CalculateGDS(clusters.DA = NULL)
}
\arguments{
\item{clusters.DA}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
