% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R,
%   R/oxy-ModulonR_functions_wo_comments.R
\name{DiscriminantAnalysis}
\alias{DiscriminantAnalysis}
\title{FUNCTION_TITLE}
\usage{
DiscriminantAnalysis(
  mat,
  method = "OPLS",
  annotation,
  BackgroundClasses = NULL,
  QueryClasses = NULL
)

DiscriminantAnalysis(
  mat,
  method = "OPLS",
  annotation,
  BackgroundClasses = NULL,
  QueryClasses = NULL
)
}
\arguments{
\item{mat}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION, Default: 'OPLS'}

\item{annotation}{PARAM_DESCRIPTION}

\item{BackgroundClasses}{PARAM_DESCRIPTION, Default: NULL}

\item{QueryClasses}{PARAM_DESCRIPTION, Default: NULL}

\item{classes}{PARAM_DESCRIPTION, Default: NULL}

\item{TargetState}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
