% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{DiscriminantAnalysis}
\alias{DiscriminantAnalysis}
\title{FUNCTION_TITLE}
\usage{
DiscriminantAnalysis(
  mat,
  method = "OPLS",
  annotation,
  classes = NULL,
  TargetState = NULL
)
}
\arguments{
\item{mat}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION, Default: 'OPLS'}

\item{annotation}{PARAM_DESCRIPTION}

\item{classes}{PARAM_DESCRIPTION, Default: NULL}

\item{TargetState}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
