% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{FindFeatureCluster}
\alias{FindFeatureCluster}
\title{FUNCTION_TITLE}
\usage{
FindFeatureCluster(
  mat,
  annotation = NULL,
  cluster.nums = 2:10,
  distance.method = DISTANCE_METHODS,
  clustering.method = CLUSTERING_METHODS,
  ...
)
}
\arguments{
\item{mat}{PARAM_DESCRIPTION}

\item{annotation}{PARAM_DESCRIPTION, Default: NULL}

\item{cluster.nums}{PARAM_DESCRIPTION, Default: 2:10}

\item{distance.method}{PARAM_DESCRIPTION, Default: DISTANCE_METHODS}

\item{clustering.method}{PARAM_DESCRIPTION, Default: CLUSTERING_METHODS}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
