% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{GetHierarchicalClusters}
\alias{GetHierarchicalClusters}
\title{FUNCTION_TITLE}
\usage{
GetHierarchicalClusters(
  mat,
  distance.method = DISTANCE_METHODS,
  clustering.method = CLUSTERING_METHODS,
  scale = TRUE,
  ...
)
}
\arguments{
\item{mat}{PARAM_DESCRIPTION}

\item{distance.method}{PARAM_DESCRIPTION, Default: DISTANCE_METHODS}

\item{clustering.method}{PARAM_DESCRIPTION, Default: CLUSTERING_METHODS}

\item{scale}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
