% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{GetSignatureAUC}
\alias{GetSignatureAUC}
\title{FUNCTION_TITLE}
\usage{
GetSignatureAUC(signature.list, mat, rankings = NULL, scale = FALSE)
}
\arguments{
\item{signature.list}{PARAM_DESCRIPTION}

\item{mat}{PARAM_DESCRIPTION}

\item{rankings}{PARAM_DESCRIPTION, Default: NULL}

\item{scale}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[AUCell]{AUCell_buildRankings}}, \code{\link[AUCell]{AUCell_calcAUC}}, \code{\link[AUCell]{aucellResults-class}}
}
