% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{ModulonIdent}
\alias{ModulonIdent}
\title{FUNCTION_TITLE}
\usage{
ModulonIdent(data, annotation, k.range = c(2, 30))
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{annotation}{PARAM_DESCRIPTION}

\item{k.range}{PARAM_DESCRIPTION, Default: c(2, 30)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
