% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{ModulonPert}
\alias{ModulonPert}
\title{Modulon Perturbation}
\usage{
ModulonPert(
  Regulons,
  Modulons,
  ExpMat,
  annotation,
  BackgroundClasses = NULL,
  TargetState,
  TargetModulon,
  CombSize,
  Weights = NULL
)
}
\arguments{
\item{Regulons}{List object with the regulon constituent elements.}

\item{Modulons}{List object with the modulon constituent elements.}

\item{ExpMat}{Matrix of gene expression with features (genes) in rows and samples (cells) in columns.}

\item{annotation}{Character vector with the class (phenotype/state) of the samples (cells).}

\item{BackgroundClasses}{Classes to compare with the query class. If NULL, all the classes in annotation except the query class will be included as background. Default: NULL.}

\item{TargetState}{Character vector with the names of the classes to be analyzed. If NULL, all the classes in annotation will be analyzed. Default: NULL.}

\item{TargetModulon}{Name of the target modulon.}

\item{CombSize}{Number of elements in the combination.}

\item{Weights}{Data frame with the weights of the regulatory influence between transcription factors and target genes. If NULL, a default weight of 1 is used for all regulations. Default: NULL.}
}
\value{
List of elements, namely, Bipartite_Graph, Combinations, and Modulon_Targets_DA:

\itemize{
\item Bipartite_Graph: Data frame with the bipartite graph between transcription factors and regulated genes.
}

\itemize{
\item Combinations: Data frame with the ranking of combinations.
}

\itemize{
\item Modulon_Targets_DA: Gene expression discriminant analysis complete results.
}
}
\description{
Wrapper function to perform the third step of a modulon analysis: modulon perturbation analysis.
}
