% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{ModulonSelect}
\alias{ModulonSelect}
\title{FUNCTION_TITLE}
\usage{
ModulonSelect(
  data,
  modulons,
  annotation,
  BackgroundClasses = NULL,
  TargetState = NULL
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{modulons}{PARAM_DESCRIPTION}

\item{annotation}{PARAM_DESCRIPTION}

\item{BackgroundClasses}{PARAM_DESCRIPTION, Default: NULL}

\item{TargetState}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
