% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonR_functions.R
\name{range01}
\alias{range01}
\title{Range \link{0,1}}
\usage{
range01(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list object with the normalized values, and the maximum and minimum values of the input vector.
}
\description{
Function to normalize absolute values of a numeric vector as follows: (x-min(x))/(max(x)-min(x)).
}
\examples{
\dontrun{
if(interactive()){
range01(c('-50','-5','5','50','100')
 }
}
}
